Adding current project files
These files are for the current paver patterns I've created.  I've also included the inspiration of the designs along with a bmp that was used in the creation of the Wave pattern.  The pavers were designed in Fusion 360 and exported as STL files.  Future updates may include links to the Fusion 360 files or the actual files themselves.